+++ esm/apprun-dev-tools.js
@@ -2,12 +2,6 @@ var _a;
 import app from './app';
 import toHTML from './vdom-to-html';
 import { _createEventTests, _createStateTests } from './apprun-dev-tools-tests';
-import yaml from 'js-yaml';
-function replacer(key, value) {
-    if (typeof value === 'function')
-        return value.toString(); // value.toString();
-    return ['', null].includes(value) || (typeof value === 'object' && (value.length === 0 || Object.keys(value).length === 0)) ? undefined : value;
-}
 function createProxy(obj) {
     const handler = {
         get(target, property, receiver) {
@@ -96,7 +90,7 @@ const viewComponents = state => {
         return app.h("li", null,
             app.h("div", null, component.constructor.name),
             app.h("div", null,
-                app.h("pre", null, htmlEncode(yaml.dump(component_def, { replacer })))),
+                app.h("pre", null, htmlEncode(JSON.stringify(component_def, null, 2)))),
             app.h("br", null));
     }));
     return app.h("ul", null, state.map(({ element, comps }) => app.h("li", null,